% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.calculate_PK_parameters.R
\name{calculate_PK_parameters}
\alias{calculate_PK_parameters}
\title{calculate_PK_parameters}
\usage{
calculate_PK_parameters(data, id_col, time_col, conc_col, format = "wider")
}
\arguments{
\item{id_col}{个体编号列}

\item{time_col}{时间列}

\item{conc_col}{血药浓度列}

\item{format}{结果为宽型数据还是长型数据(long)，默认为宽型数据(wider)}

\item{df}{原始数据}
}
\value{
df data.frame
}
\description{
函数作用：此函数快捷计算AUC、Cmax、Cmin、CLast
}
\examples{
#构建示例数据
data<-data.frame(
ID = c(1,1,2,2),
TIME = c(1, 2, 1, 2),
IPRED = c(10, 15, 5, 8))
#示例代码
pk_results <- calculate_PK_parameters(data, "ID", "TIME", "IPRED")
print(pk_results)
}
