% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_htwt.R
\name{clean_htwt}
\alias{clean_htwt}
\title{Height and Weight anomaly detection.}
\usage{
clean_htwt(df, lower_percentile)
}
\arguments{
\item{df}{Data frame containing specific columns:

\strong{HT:} Patient height (cm).

\strong{WT:} Patient weight (kg).

\strong{SEX:} Patient sex 0 = female, 1 = female.

either

\strong{TIME:} Dates when corresponding height and weight values
were collected. Must be R datetime object.

\strong{BIRTH_DT:} Birth date of patient. Must be R date or
datetime object.

or

\strong{AGE_M:} Age of patient in months.}

\item{lower_percentile}{Integer value representing the lower
percentile, upper percentile will automatically be calculated.

Options: 3,5,10,25

Example: lower_percentile = 3. Normal will be defined as inbetween
fith and ninety fifth percentile (3-97).

lower_percentile = 5, (5-95).

lower_percentile = 10, (10-90).

lower_percentile = 25, (25-75).}
}
\value{
Returns a data frame identifying height weight anomalies
with specific colums:

\strong{HT_FLAG:} Numerical value. >0 signifying distance from
normal if height is abnormal. 0 if there is no abnormality in
height detected.

\strong{WT_FLAG:} Numerical value. >0 signifying distance from
normal if weight is abnormal. 0 if there is no abnormality in
weight detected.

\strong{AGE_Y:} Age of patient in years.
}
\description{
Uses CDC height and weight expected values to determine anomalies
in height and weight values for cleaning purposes.
}
\details{
See \url{https://www.cdc.gov/growthcharts/percentile_data_files.htm}
for data sources used to detect anomalies.

Datasets are included in package as statage, lenageinf, wtage,
and wtageinf.
}
\examples{
clean_htwt(htwt_example, 3)
}
\keyword{clean}
\keyword{cleaning,}
\keyword{height,}
\keyword{weight,}
