% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_three_Quartiles_value.R
\name{convert_three_Quartiles_value}
\alias{convert_three_Quartiles_value}
\title{convert_three_Quartiles_value}
\usage{
convert_three_Quartiles_value(
  df,
  column_names,
  suffix = "_quartiles",
  label_format = "Q_numeric"
)
}
\arguments{
\item{df}{需要转换的数据框}

\item{column_names}{需要转换的列,可以为多列}

\item{suffix}{新形成分组列标签后缀}

\item{label_format}{区间展示形式，可以从Q、Q_numeric、numeric中选择，默认为Q_numeric}
}
\value{
df data.frame
}
\description{
函数作用：此函数是将数据框中的多列批量转化为三分位数区间，分组列按照数值区间表示，例如将Cmax按照三分卫数区间分组
}
\examples{
#构建示例数据
df <- data.frame(
value1 = c(5, 12, 8, 15, 7),
value2 = c(18, 22, 25, 10, 30)
)
#对指定列进行四分位数转换
df <- convert_three_Quartiles_value(df, column_names=c("value1", "value2"))
}
