% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.geom_boxplot_y_vs_multiple_x.R
\name{geom_boxplot_y_vs_multiple_x}
\alias{geom_boxplot_y_vs_multiple_x}
\title{geom_boxplot_y_vs_multiple_x}
\usage{
geom_boxplot_y_vs_multiple_x(
  data,
  x_vars,
  y_var,
  y_labels,
  x_labels,
  save_path,
  width1 = 10,
  height1 = 10,
  sample_size,
  P,
  show_title
)
}
\arguments{
\item{data}{原始数据}

\item{x_vars}{X轴变量,为分类变量}

\item{y_var}{Y轴变量，为连续性变量}

\item{y_labels}{定义Y轴标题}

\item{x_labels}{定义X轴标题}

\item{save_path}{定义图片保存路径}

\item{width1}{输出图片宽度}

\item{height1}{输出图片高度}

\item{sample_size}{默认不展示样本量，若展示sample_size=TRUE}

\item{P}{是否展示P值，若展示，填入P=TRUE}

\item{show_title}{是否展示图标题,默认为无图标题,若展示show_title=TRUE}
}
\description{
函数作用：绘制多个X变量与某个Y变量的箱型图,输出到特定路径
}
\examples{
# 示例数据
data <- data.frame(
C_avg = rep(c("[5.2-43.1]", "[37.1-53.5]", "[53.5-97]", "[97-296]"), each = 4),
Response1 = rnorm(16, mean = 50, sd = 10),
Response1 = rnorm(16, mean = 50, sd = 10)
)
#示例代码
geom_boxplot_y_vs_multiple_x(data=data,
x_vars=c("C_avg"),
y_var=c("Response1")
y_labels="Response1",
x_labels=c("C_avg(μg/ml)")
save_path="C:/Users/haiyun.yang/Desktop/111"
)
}
