% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.group_by_threshold.R
\name{group_by_threshold}
\alias{group_by_threshold}
\title{data manipulation}
\usage{
group_by_threshold(data, columns, thresholds)
}
\arguments{
\item{data}{原始数据}

\item{columns}{需要转换的列}

\item{thresholds}{界值}
}
\value{
data data.frame
}
\description{
#' 函数作用：比如将人群年龄按照是否大于等于65岁分组，此函数可以快捷实现，
只需要输入列名和界值
}
\examples{
df <- data.frame(
value1 = c(5, 12, 8, 10, 7),
value2 = c(18, 22, 25, 20, 30))
使用group_by_threshold函数进行分组
df_grouped <- group_by_threshold(df, c("value1", "value2"), c(10, 20))
查看结果
print(df_grouped)
}
