% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{htwt_example}
\alias{htwt_example}
\title{Height and Weight example dataset.}
\format{A data frame with 399 rows and 6 variables:
\describe{
  \item{TIME}{Time that the height and weight were recorded.}
  \item{HT}{Height in cm}
  \item{WT}{Weight in kg}
  \item{BIRTH_DT}{Birth date}
  \item{SEX}{0 is female, 1 is male}
  \item{AGE_Y}{Age in years}
}}
\usage{
htwt_example
}
\description{
Ages between 0 and 20 years.
}
\keyword{datasets}
