% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_plots.R
\name{individual_plots}
\alias{individual_plots}
\title{Individual prediction plots.}
\usage{
individual_plots(df, filename, ID = ID, TIME = TIME, DV = DV,
  PRED = PRED)
}
\arguments{
\item{df}{Data frame containing ID, height, the time weight and
height were collected, sex, and eighter birth date or age in months.}

\item{filename}{Name of file to save pdf output.}

\item{ID}{Column name for unique individual identifier.}

\item{TIME}{Column name for datetime or time when concentration
values were collected.}

\item{DV}{Column name for observed concentration values.}

\item{PRED}{Column name for predicted concentration values.}
}
\value{
A pdf containing observed vs predicted and time vs observed
and predicted plots for each ID. The pdf will be saved
as 'filename' in the working directory.
}
\description{
Plots individual observed vs predicted values as a diagnostic
tool for pk/pd modeling.
}
\details{
If any dataframe input column has a different name, define
that columns name in the input parameters (See Examples).
See http://r4ds.had.co.nz chapters 11.4, 11.5 and 16 for creating
date or datetime R objects.
}
\examples{
individual_plots(pred_example, "pred_example_plots.pdf")
# pred_example_plots.pdf should now be in your working directory
}
\keyword{individual}
\keyword{plot,}
\keyword{prediction,}
