% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lenageinf}
\alias{lenageinf}
\title{Length based on age (0m-36m).}
\format{A data frame with 74 rows and 14 variables:
\describe{
  \item{Sex}{1 = male, 2 = female}
  \item{Agemos}{Age in months}
  \item{L}{Use to find new z-score values (see details below)}
  \item{M}{Use to find new z-score values (see details below)}
  \item{S}{Use to find new z-score values (see details below)}
  \item{P3}{Third percentile length in cm}
  \item{P5}{Fifth percentile length in cm}
  \item{P10}{Tenth percentile length in cm}
  \item{P25}{Twenty fifth percentile length in cm}
  \item{P50}{Fiftieth percentile length in cm}
  \item{P75}{Seventy fifth percentile length in cm}
  \item{P90}{Ninetieth percentile length in cm}
  \item{P95}{Ninety fifth percentile length in cm}
  \item{P97}{Ninety seventh percentile length in cm}
}}
\source{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm}
}
\usage{
lenageinf
}
\description{
Length value by percentiles based on sex and age (0 to 36 months).
}
\details{
Contains L, M, and S parameters to determine new z-scores from:

X = M * (1 + L*S*Z)^(1 / L), L != 0

X = M * exp(S * Z), L == 0
}
\keyword{datasets}
