% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10.logistic_plot_y_vs_multiple_x.R
\name{logistic_plot_y_vs_multiple_x}
\alias{logistic_plot_y_vs_multiple_x}
\title{logistic_plot_y_vs_multiple_x}
\usage{
logistic_plot_y_vs_multiple_x(
  data,
  x_vars,
  y_var,
  y_labels,
  x_labels = NULL,
  save_path = NULL,
  DOSE,
  width1 = 7,
  height1 = 5,
  show_title
)
}
\arguments{
\item{data}{data.franme}

\item{x_vars}{多个暴露量}

\item{y_var}{结局指标}

\item{y_labels}{Y轴标题}

\item{x_labels}{多个X轴标题}

\item{save_path}{逻辑回归图保存路径}

\item{DOSE}{是否将暴露箱型图分组展示，可以不填}

\item{width1}{输出图片宽度}

\item{height1}{输出图片高度}

\item{show_title}{是否展示图标题,默认为无图标题,若展示show_title=TRUE}
}
\description{
此函数将旨在快捷进行逻辑回归拟合图，比如暴露-TEAE的图，可以批量绘制
}
\examples{
#构建示例数据
data <- data.frame(
status = rbinom(100, 1, 0.5),
Humidity = rnorm(100, mean = 50, sd = 10),
DOSE=rep(c("[5.2-43.1]", "[37.1-53.5]", "[53.5-97]", "[97-296]"), each = 25),
Wind_Speed = rnorm(100, mean = 15, sd = 5))
data$status <- as.numeric(data$status)
data$DOSE <- as.factor( data$DOSE)
#示例运行代码
logistic_plot_y_vs_multiple_x (data, y_var="status",
 x_vars=c("Humidity", "Wind_Speed"),
 x_labels=c("A", "Air Humidity"),
 y_labels="TEAE",
 save_path="C:/Users/haiyun.yang/Desktop/222")
}
