% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20.NCA.R
\name{NCA}
\alias{NCA}
\title{函数作用：实现NCA分析}
\usage{
NCA(data, ID, TIME, DV, DOSE, adminType = "extravascular", Infusion_duration)
}
\arguments{
\item{data}{原始数据}

\item{ID}{个体编号列}

\item{TIME}{时间列}

\item{DV}{血药浓度列}

\item{adminType}{给药方式,可以从"extravascular","iv-infusion","iv-bolus"选择一种，若为静脉输注需要填写输注时间::Infusion_duration=XX}
}
\value{
df data.frame
}
\description{
函数作用：实现NCA分析
}
\examples{
#构建示例数据
data <- data.frame(
ID=1,
TIME = c(0,0.25,0.5,1,1.5,2,3,4,6,8,12,16,24),
DOSE=100,
DV=c(0, 0.07, 0.14, 0.21, 0.24, 0.27, 0.26, 0.25, 0.22, 0.19, 0.13, 0.081, 0.033)
)
#示例代码
NAC <- NCA(data=data,TIME = "TIME",ID = "ID",DV="DV",adminType="extravascular")
}
