% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PFS_survival_by_categories.R
\name{plot_PFS_survival_by_categories}
\alias{plot_PFS_survival_by_categories}
\title{plot_PFS_survival_by_categories}
\usage{
plot_PFS_survival_by_categories(
  data,
  time_col,
  event_col,
  categorical_vars,
  output_dir,
  legend_titles
)
}
\arguments{
\item{data}{原始数据}

\item{time_col}{时间列}

\item{event_col}{删失列}

\item{categorical_vars}{需要研究的暴露、协变量分组列}

\item{output_dir}{保存路径}

\item{legend_titles}{图例标题}
}
\description{
函数作用：此函数是可以批量绘制PFS的KM曲线
}
\examples{
#构建示例数据
# 示例使用
data <- data.frame(
time = rexp(200, rate = 0.1),
status = rbinom(200, 1, 0.7),
group = sample(c("A", "B"), 200, replace = TRUE),
treatment = sample(c("Control", "Treatment"), 200, replace = TRUE))
#示例代码
plot_PFS_survival_by_categories(data,time_col="time",event_col="status",output_dir="",legend_titles=c("A","B"))
}
