% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_response_percentage.R
\name{plot_response_percentage}
\alias{plot_response_percentage}
\title{plot_response_percentage}
\usage{
plot_response_percentage(
  data,
  group_column,
  response_column,
  target_value,
  chart_title,
  x_label,
  y_label,
  file_path,
  show_text
)
}
\arguments{
\item{data}{原始数据}

\item{group_column}{X轴变量,为分类变量}

\item{response_column}{Y轴变量，为分类变量}

\item{target_value}{定义统计Y轴变量的哪个分类}

\item{chart_title}{图片标题}

\item{x_label}{定义X轴标题}

\item{y_label}{定义Y轴标题}

\item{file_path}{定义图片保存路径}

\item{show_text}{是否展示2/4(50%)的文字}
}
\description{
函数作用：以ADA阴阳性为例，此函数可以快捷统计ADA阳性在某分组中的百分比例,
绘制柱状图，并输出到特定文件夹
}
\examples{
# 示例数据
data <- data.frame(
C_avg = rep(c("[5.2-43.1]", "[37.1-53.5]", "[53.5-97]", "[97-296]"), each = 4),
Response1 = c(1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0),
Response2 = c(0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1)
)
 #示例代码
 plot_response_percentage(data, "C_avg", "Response1",
 target_value, "Clinical Remission per C_avg Interval",
 "C_avg Interval", "\% Clinical Remission",
 "C:/Users/haiyun.yang/Desktop/111/plot.png")
}
