% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.plot_stacked_categories.R
\name{plot_stacked_categories}
\alias{plot_stacked_categories}
\title{plot_stacked_categories}
\usage{
plot_stacked_categories(
  data,
  x_vars,
  y_var,
  y_labels,
  x_labels = NULL,
  save_path = NULL,
  width1 = 7,
  height1 = 5
)
}
\arguments{
\item{data}{data.franme}

\item{x_vars}{分组变量}

\item{y_var}{结局指标}

\item{y_labels}{Y轴标题}

\item{x_labels}{X轴标题}

\item{save_path}{逻辑回归图保存路径}

\item{width1}{输出图片宽度}

\item{height1}{输出图片高度}
}
\description{
此函数将旨在绘制堆积柱状图
}
\examples{
#构建示例数据
data<- data.frame(
Group = rep(c("A", "B", "C"), each = 10),
Group1 = rep(c("A", "B", "C"), each = 10),
Category = sample(c("Pass", "Fail"), 30, replace = TRUE, prob = c(0.7, 0.3))
)
#示例代码
plot_stacked_categories(data, x_vars=c"Group",
y_var="Category", y_labels="zzz",x_labels=c"BBB",
save_path="C:/Users/haiyun.yang/Desktop/222",
width1=8, height1=5)
}
