% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pred_example}
\alias{pred_example}
\title{Predicted and observed example dataset.}
\format{A data frame with 556 rows and 4 variables:
\describe{
  \item{ID}{Individual identifier}
  \item{TIME}{Time that the observed concentrations were recorded}
  \item{DV}{Observed concentration values}
  \item{PRED}{Predicted concentration values}
}}
\usage{
pred_example
}
\description{
Predicted and observed example dataset.
}
\keyword{datasets}
