% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_logistic_regression(1).R
\name{run_logistic_regression}
\alias{run_logistic_regression}
\title{run_logistic_regression}
\usage{
run_logistic_regression(
  data,
  outcome_var,
  predictor_vars,
  output_dir,
  xlim1 = c(0, 5),
  ticks_at1 = c(0, 5),
  multi_predictor_vars,
  width1 = 10,
  height1 = 10
)
}
\arguments{
\item{data}{data.franme}

\item{outcome_var}{结局指标，必须是因子}

\item{predictor_vars}{需要分析的协变量}

\item{output_dir}{结果输出路径}

\item{xlim1}{森林图X轴刻度范围}

\item{ticks_at1}{森林图X轴刻度线}

\item{multi_predictor_vars}{填入基于逐步法筛选得到的协变量}

\item{width1}{输出的森林图宽度}

\item{height1}{输出的森林图高度}
}
\description{
run_logistic_regression
}
\details{
版权作者:杨海云
未禁许可，禁止外传
}
\examples{
#构建示例数据
set.seed(123)
data <- data.frame(
status = rbinom(100, 1, 0.5),  # 结局事件
age = rnorm(100, 50, 10),      # 年龄
sex = rbinom(100, 1, 0.5),     # 性别
treatment = rbinom(100, 1, 0.6) # 治疗组别
)
#以下为示例代码
 run_logistic_regression(data=data, outcome_var="status",
 predictor_vars=c("age", "sex", "treatment"),
 output_dir="C:/Users/haiyun.yang/Desktop/222",xlim1=c(0,5),ticks_at1=c(0,5)
 ,multi_predictor_vars=c("age"))
}
