% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wtage}
\alias{wtage}
\title{Weight based on age (2y-20y).}
\format{A data frame with 436 rows and 14 variables:
\describe{
  \item{Sex}{1 = male, 2 = female}
  \item{Agemos}{Age in months}
  \item{L}{Use to find new z-score values (see details below)}
  \item{M}{Use to find new z-score values (see details below)}
  \item{S}{Use to find new z-score values (see details below)}
  \item{P3}{Third percentile weight in kg}
  \item{P5}{Fifth percentile weight in kg}
  \item{P10}{Tenth percentile weight in kg}
  \item{P25}{Twenty fifth percentile weight in kg}
  \item{P50}{Fiftieth percentile weight in kg}
  \item{P75}{Seventy fifth percentile weight in kg}
  \item{P90}{Ninetieth percentile weight in kg}
  \item{P95}{Ninety fifth percentile weight in kg}
  \item{P97}{Ninety seventh percentile weight in kg}
}}
\source{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm}
}
\usage{
wtage
}
\description{
Weight values by percentiles based on sex and age (2 to 20 years).
}
\details{
Contains L, M, and S parameters to determine new z-scores from:

X = M * (1 + L*S*Z)^(1 / L), L != 0

X = M * exp(S * Z), L == 0
}
\keyword{datasets}
