% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wtstat}
\alias{wtstat}
\title{Weight based on stature (77cm-121cm).}
\format{A data frame with 92 rows and 15 variables:
\describe{
  \item{Sex}{1 = male, 2 = female}
  \item{Height}{Height in cm}
  \item{L}{Use to find new z-score values (see details below)}
  \item{M}{Use to find new z-score values (see details below)}
  \item{S}{Use to find new z-score values (see details below)}
  \item{P3}{Third percentile weight in kg}
  \item{P5}{Fifth percentile weight in kg}
  \item{P10}{Tenth percentile weight in kg}
  \item{P25}{Twenty fifth percentile weight in kg}
  \item{P50}{Fiftieth percentile weight in kg}
  \item{P75}{Seventy fifth percentile weight in kg}
  \item{P85}{Eighty fifth percentile weight in kg}
  \item{P90}{Ninetieth percentile weight in kg}
  \item{P95}{Ninety fifth percentile weight in kg}
  \item{P97}{Ninety seventh percentile weight in kg}
}}
\source{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm}
}
\usage{
wtstat
}
\description{
Weight values by percentiles based on sex and height (77 to 121 cm).
}
\details{
Contains L, M, and S parameters to determine new z-scores from:

X = M * (1 + L*S*Z)^(1 / L), L != 0

X = M * exp(S * Z), L == 0
}
\keyword{datasets}
